+++ lib/Dialect/TTIR/Transforms/TTIRFusing.cpp
@@ -293,7 +293,7 @@ private:
 
     mlir::func::FuncOp funcOp = conv2dOp->getParentOfType<mlir::func::FuncOp>();
     llvm::SmallPtrSet<BlockArgument, 4> constParams =
-        ttmlir::utils::populateConstParams(funcOp);
+        mlir::tt::populateConstParams(funcOp);
     auto isConstant = [&constParams, conv2dOp](mlir::Value value) {
       if (auto blockArg = mlir::dyn_cast<BlockArgument>(value)) {
         return constParams.contains(blockArg);
